Add option for displaying and counting proper full coverage
We need to do two things for that: First is to create empty tests, and then to take coverage report, run go cov on it
and take the full coverage there.

For the first part - creating empty tests - it's a little tricky.

* first we list all packages, to get file paths/names
* then we get all the packages without tests, through go test list
* then we create dummy test files
* then we run the gotestiful
* then we remove the dummy files

We need to do some trickery to still print the originally no-test files
as skipped.

For the second part - we need to save the go coverage file always, but in the case user doesn't want it, delete it again.