Add option for including files with empty tests in coverage reports
The way we do this is a little tricky.
* first we list all packages, to get file paths/names
* then we get all the packages without tests, through go test list
* then we create dummy test files
* then we run the gotestiful
* then we remove the dummy files

We need to do some trickery to still print the originally no-test files
as skipped.

Note that with the current coverage reporting, which is wrong (as it just averages all coverages),
there is no difference in overall coverage reporting; I plan to do one more PR that will fix the overall coverage
reporting