fix: remove log.Fatal in logic and replace with Errors (#5)
One negative of log.Fatal() is that, unlike panic or error, it does *not* run defered
functions. It immediately exits the program and all other goroutines that are running.

It doesn't currently matter, as we don't do any cleanup in any defer function, but it can be confusing.

Also it's cleaner, I think.